pyproject.toml
[project]
name = "integrator-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.8",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "langchain>=1.2.17",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.2.1",
    "langfuse>=4.5.1",
    "openai>=2.33.0",
    "python-dotenv>=1.2.2",
    "tiktoken>=0.12.0",
]
